feat 権限レベル変更コマンド追加
